% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximatePosterior.R
\name{approximatePosterior}
\alias{approximatePosterior}
\title{Approximate posterior using Bayesian model averaging}
\usage{
approximatePosterior(
  output_dir = ".",
  data = NULL,
  group = NULL,
  hyper_mu = c(1, 0.5, 0.25, 0.2, 0.1, 0.25),
  hyper_prec = c(400, 400, 20, 800, 800, 20),
  alpha_marg = c(1.625, 1),
  alpha_cop = c(1, 1, 1, 1),
  mcmc_niter = 3500,
  mcmc_nburnin = 1000,
  mcmc_nthin = 1,
  output_prefix = "bma_post",
  overwrite = FALSE
)
}
\arguments{
\item{output_dir}{Directory where the scripts should be written.}

\item{data}{Matrix with two columns (one column per biomarker).}

\item{group}{a vector of group assignments (length must be the column length of \code{data}).}

\item{hyper_mu}{Hyper-mean parameters for data analysis.}

\item{hyper_prec}{Hyper-precision parameters for data analysis.}

\item{alpha_marg}{Dirichlet prior parameters for the two marginal model probabilities.}

\item{alpha_cop}{Dirichlet prior parameters for the four copula models.}

\item{mcmc_niter}{Number of MCMC iterations per chain (16 chains used).}

\item{mcmc_nburnin}{Number of burn-in iterations.}

\item{mcmc_nthin}{Level of thinning for each MCMC chain.}

\item{output_prefix}{Prefix used for output CSV files.}

\item{overwrite}{Logical; if \code{FALSE}, existing files will not be overwritten.}
}
\value{
Returns a list with the posterior sample, R hat statistics, and posterior probabilities for
candidate model selection; the posterior sample may be output to a .csv file
}
\description{
This function takes in a bivariate sample, prior distributions, and
parameters for various MCMC settings. The function outputs a data frame
with a sample from the posterior distribution, R hat statistics to
assess Markov chain convergence, and summaries of posterior probabilities
for candidate model selection
}
\details{
It may take a few minutes for the MCMC model to compile using NIMBLE. Once compiled,
a progress bar will populate in the console to display progress for each MCMC chain.
}
\examples{
\dontrun{
## get example data set
data.temp <- cbind(rgamma(18, 2, 1), rgamma(18, 2, 1))

## save posterior approximation results
approximatePosterior(data = data.temp,
       group = rep(1:4, c(6, 4, 4, 4)),
       hyper_mu = c(1, 0.5, 0.25, 0.2, 0.1, 0.25),
}

}
