% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximatePosterior.R
\name{bma.sample-methods}
\alias{bma.sample-methods}
\alias{print.bma.sample}
\title{Printing summaries of posterior approximations}
\usage{
\method{print}{bma.sample}(x, ...)
}
\arguments{
\item{x}{a bma.sample object}
}
\description{
Helper function to print bma.sample objects
}
\keyword{internal}
