% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{contourDoses}
\alias{contourDoses}
\title{Contour plots of the posterior predictive distribution}
\usage{
contourDoses(
  output_dir = ".",
  bma_post = NULL,
  n_grid = 300,
  n_samp = 25000,
  n_cores = max(1, parallel::detectCores() - 1),
  output_prefix = "contour",
  overwrite = TRUE
)
}
\arguments{
\item{output_dir}{Directory where the scripts should be written.}

\item{bma_post}{Output saved from a previous \code{approximatePosterior()} function call.}

\item{n_grid}{Number of grid points in each dimension for the contour plot (at most 500).}

\item{n_samp}{Number of observations simulated from the posterior predictive distribution for plotting}

\item{n_cores}{Number of parallel cores to use with parallelization.}

\item{output_prefix}{Prefix used for output PDF and CSV files.}

\item{overwrite}{Logical; if \code{FALSE}, existing CSV files will not be overwritten.}
}
\value{
Returns a contour plot for each dose as output to a PDF file; the samples from
the posterior predictive distribution used to make the contour plot for each dose may
be output to CSV files. Those samples can be used to create more tailored contour plots.
}
\description{
This function takes in the output from the function \code{approximatePosterior()}.
The function outputs contour plots for each dose with the data input into
\code{approximatePosterior()} superimposed.
}
\examples{
\dontrun{
## save posterior approximation results
bma.post.temp <- approximatePosterior(...)

## obtain contour plots
contourDoses(bma_post = bma.post.temp,
             n_grid = 300,
             n_samp = 25000,
             n_cores = max(1, parallel::detectCores() - 1),
             output_prefix = "contour",
             overwrite = TRUE)
}

}
