% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGen.R
\name{dataGen}
\alias{dataGen}
\title{Generate a (hierarchical) sample across various doses}
\usage{
dataGen(
  n = 18,
  gs = c(6, 4, 4, 4),
  marg1,
  marg2,
  cop,
  hier = TRUE,
  hyper_mu,
  hyper_prec,
  mus_nohier = NULL,
  sigmas_nohier = NULL,
  seed = 1
)
}
\arguments{
\item{n}{sample size}

\item{gs}{a vector of group sizes}

\item{marg1}{The marginal distribution for endpoint 1 (1 = gamma, 2 = lognormal)}

\item{marg2}{The marginal distribution for endpoint 2 (1 = gamma, 2 = lognormal)}

\item{cop}{The copula model (1 = independence, 2 = Clayton, 3 = Gaussian, 4 = Gumbel)}

\item{hier}{boolean to denote that group-specific parameters are not provided}

\item{hyper_mu}{mean hyperparameters used for data generation. The first three inputs are
for the \eqn{\mu} parameters of \eqn{m_1} and \eqn{m_2}, \eqn{c_1} and \eqn{c_2}, and logit(\eqn{\tau}).
The second three inputs are for \eqn{\sigma} parameters of \eqn{m_1} and \eqn{m_2},
\eqn{c_1} and \eqn{c_2}, and logit(\eqn{\tau}).}

\item{hyper_prec}{precision hyperparameters used for data generation. The order of the parameters
is the same as \code{hyper_mu}}

\item{mus_nohier}{mu parameters for each group (\code{hier = FALSE}). These are the mean parameters
for \eqn{m_1}, \eqn{m_2}, \eqn{c_1}, \eqn{c_2}, and logit(\eqn{\tau}) (in order)}

\item{sigmas_nohier}{sigma parameters for each group (\code{hier = FALSE}). These are the standard
deviation parameters for \eqn{m_1}, \eqn{m_2}, \eqn{c_1}, \eqn{c_2}, and logit(\eqn{\tau}) (in order)}

\item{seed}{an integer seed used for reproducibility}
}
\value{
A list of length 3 for use in NIMBLE. Element 1 contains the number of participants, groups, and dose assignments. Element 2 contains the bivariate relative increases. Element 3 contains the data-generating parameters (returned as a check).
}
\description{
This function takes in a sample size of n as well as inputs for the
group sizes and hierarchical structure and returns a sample of n
bivariate observations
}
\examples{
# generate a sample of size 18 across 4 doses using a hierarchical model
dataGen(n = 18, gs = c(6, 4, 4, 4), marg1 = 1, marg2 = 1,
        cop = 1, hyper_mu = c(1, 0.5, 0.25, 0.2, 0.1, 0.25),
        hyper_prec = c(400, 400, 20, 800, 800, 20), seed = 1)

}
