% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initFromData}
\alias{initFromData}
\title{Helper function to get initial values for the means and CVs for each dose}
\usage{
initFromData(y, group)

initFromData(y, group)
}
\arguments{
\item{y}{a vector of observations (one endpoint)}

\item{group}{a vector denoting group/dose membership}
}
\value{
A list of hyperparameters

A list of hyperparameters
}
\description{
This function takes in the data corresponding to one endpoint and
returns a list of hyperparameters for MCMC initialization

This function takes in the data corresponding to one endpoint and
returns a list of hyperparameters for MCMC initialization. This function
is only used internally.
}
\keyword{internal}
