% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorPredictive.R
\name{post.pred-methods}
\alias{post.pred-methods}
\alias{print.post.pred}
\title{Printing summaries of posterior approximations}
\usage{
\method{print}{post.pred}(x, ...)
}
\arguments{
\item{x}{a post.pred object}
}
\description{
Helper function to print post.pred objects
}
\keyword{internal}
