% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorPredictive.R
\name{posteriorPredictive}
\alias{posteriorPredictive}
\title{Posterior predictive inference using Bayesian model averaging}
\usage{
posteriorPredictive(
  output_dir = ".",
  bma_post = NULL,
  inner_rep = 500,
  bounds1 = cbind(c(0, 0, 1, 1), c(1, 1, Inf, Inf)),
  bounds2 = cbind(c(0, 1, 0, 1), c(1, Inf, 1, Inf)),
  n_cores = max(1, parallel::detectCores() - 1),
  credible = 0.95,
  output_prefix = "post_pred",
  overwrite = TRUE
)
}
\arguments{
\item{output_dir}{Directory where the scripts should be written.}

\item{bma_post}{Output saved from a previous \code{approximatePosterior()} function call.}

\item{inner_rep}{Number of samples per posterior draw.}

\item{bounds1}{A two-column matrix of lower and upper bounds for the first
endpoint that define joint events. The first column is for lower bounds, and the
second one is for upper bounds. Use \code{0}/\code{Inf} if there is no lower or upper bound.
These lower and upper bounds should not account for inversion of the first endpoint.}

\item{bounds2}{A two-column matrix of lower and upper bounds for the second
endpoint that define joint events. The rows must correspond to the rows from \code{bounds1}. The
joint event in row \eqn{k} of the matrices is such that \eqn{\tilde{Y}_1 \in} \code{c(bounds1[k, 1], bounds1[k, 2])}
and \eqn{\tilde{Y}_2 \in} \code{c(bounds2[k, 1], bounds2[k, 2])}.}

\item{n_cores}{Number of parallel cores to use with parallelization.}

\item{credible}{credibility level for the credible intervals (0.95 by default).}

\item{output_prefix}{Prefix used for output CSV files.}

\item{overwrite}{Logical; if \code{FALSE}, existing files will not be overwritten.}
}
\value{
Returns a table with posterior predictive probabilities and credible intervals
for each dose; this table may be output to a .csv file
}
\description{
This function takes in the output from the function \code{approximatePosterior()}.
The function outputs a table that summarizes the posterior predictive
probabilities for the various events input into the function with corresponding
credible intervals.
}
\examples{
\dontrun{
## save posterior approximation results
bma.post.temp <- approximatePosterior(...)

## obtain posterior predictive probabilities
posteriorPredictive(bma_post = bma.post.temp,
                    inner_rep = 500,
                    bounds1 = cbind(c(0, 0, 1, 1), c(1, 1, Inf, Inf)),
                    bounds2 = cbind(c(0, 1, 0, 1), c(1, Inf, 1, Inf)),
                    n_cores = max(1, parallel::detectCores() - 1),
                    credible = 0.95,
                    output_prefix = "post_pred",
                    overwrite = TRUE)
}

}
