% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximatePosterior.R, R/dataGen.R
\name{rBivarCop}
\alias{rBivarCop}
\title{Generate observations from the bivariate distribution}
\usage{
rBivarCop(n, marg1, m1, c1, marg2, m2, c2, cop, tau)

rBivarCop(n, marg1, m1, c1, marg2, m2, c2, cop, tau)
}
\arguments{
\item{n}{sample size}

\item{marg1}{The marginal distribution for endpoint 1 (1 = gamma, 2 = lognormal)}

\item{m1}{The mean of the marginal distribution for endpoint 1}

\item{c1}{The coefficient of variation of the marginal distribution for endpoint 1}

\item{marg2}{The marginal distribution for endpoint 2 (1 = gamma, 2 = lognormal)}

\item{m2}{The mean of the marginal distribution for endpoint 2}

\item{c2}{The coefficient of variation of the marginal distribution for endpoint 2}

\item{cop}{The copula model (1 = independence, 2 = Clayton, 3 = Gaussian, 4 = Gumbel)}

\item{tau}{Kendall's tau (must be between 0 and 1)}
}
\value{
Bivariate observations corresponding to the same dose (each column corresponds to an endpoint)

Bivariate observations corresponding to the same dose (each column corresponds to an endpoint)
}
\description{
This function takes in a sample size of \code{n} as well as inputs for the
marginal distributions, means, variances, copula, and Kendall's tau
and returns bivariate observations corresponding to the same dose

This function takes in a sample size of \code{n} as well as inputs for the
marginal distributions, means, variances, copula, and Kendall's tau
and returns bivariate observations corresponding to the same dose
}
\examples{
# generate a sample of size 5 from a single dose
rBivarCop(n = 5, marg1 = 1, m1 = 1, c1 = 0.5, marg2 = 2, m2 = 1.2, c2 = 0.65,
          cop = 2, tau = 0.5)

# generate a sample of size 5 from a single dose
rBivarCop(n = 5, marg1 = 1, m1 = 1, c1 = 0.5, marg2 = 2, m2 = 1.2, c2 = 0.65,
          cop = 2, tau = 0.5)

}
