% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBivarCopula.R
\name{rBivarCopula}
\alias{rBivarCopula}
\title{Generate from the bivariate distribution (used only by nimble)}
\usage{
rBivarCopula(n, marg1, m1, c1, marg2, m2, c2, cop, tau)
}
\arguments{
\item{n}{sample size (must be 1).}

\item{marg1}{The marginal distribution for endpoint 1 (1 = gamma, 2 = lognormal).}

\item{m1}{The mean of the marginal distribution for endpoint 1.}

\item{c1}{The coefficient of variation of the marginal distribution for endpoint 1.}

\item{marg2}{The marginal distribution for endpoint 2 (1 = gamma, 2 = lognormal).}

\item{m2}{The mean of the marginal distribution for endpoint 2.}

\item{c2}{The coefficient of variation of the marginal distribution for endpoint 2.}

\item{cop}{The copula model (1 = independence, 2 = Clayton, 3 = Gaussian, 4 = Gumbel).}

\item{tau}{The value for Kendall's tau (must be between 0 and 1 for this implementation).}
}
\value{
A bivariate observation
}
\description{
This function takes in a sample size of \code{n = 1} as well as inputs for the
marginal distributions, means, coefficients of variation, copula, and Kendall's tau
and returns a bivariate observation \code{x}
}
