% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_bivar_copula_distribution}
\alias{register_bivar_copula_distribution}
\title{Register the custom bivariate copula distribution for NIMBLE}
\usage{
register_bivar_copula_distribution()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
This function registers the custom \code{dBivarCopula} distribution so that it
can be used inside NIMBLE model code.
}
\examples{
register_bivar_copula_distribution()

}
