% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{registerBivarCopulaDistribution}
\alias{registerBivarCopulaDistribution}
\title{Register the custom bivariate copula distribution for NIMBLE}
\usage{
registerBivarCopulaDistribution()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
This function registers the custom \code{dBivarCopula} distribution so that it
can be used inside NIMBLE model code.
}
\examples{
registerBivarCopulaDistribution()

}
