% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_model_probs}
\alias{summarize_model_probs}
\title{Determine probability of each submodel having largest posterior probability}
\usage{
summarize_model_probs(mat)
}
\arguments{
\item{mat}{a matrix of posterior probabilities}
}
\value{
A vector of probabilities
}
\description{
This function takes in a matrix of posterior probabilities, with one column
per submodel and returns the probability of each submodel having the largest
posterior probability
}
